Update JPA ORM Diagnostics Trace to utilize function for Introspector update
