+++ dev/com.ibm.ws.jpa.container.core/src/com/ibm/ws/jpa/management/JPAScopeInfo.java
@@ -83,9 +83,15 @@ public class JPAScopeInfo {
                 pxmlInfo = new JPAPxmlInfo(this, pxmlRootURL); // F743-23167
                 pxmlsInfo.put(pxmlInfoKey, pxmlInfo);
 
-                // read in all the persistence unit define in this persistence.xml
-                pxmlInfo.extractPersistenceUnits(pxml); //PK62950
+                JPAIntrospection.beginPXmlInfoVisit(pxmlInfo);
+                try {
+                    // read in all the persistence unit define in this persistence.xml
+                    pxmlInfo.extractPersistenceUnits(pxml); //PK62950
+                } finally {
+                    JPAIntrospection.endPXmlInfoVisit();
+                }
             }
+
         }
 
         if (isTraceOn && tc.isEntryEnabled())
@@ -247,17 +253,15 @@ public class JPAScopeInfo {
             pxmlsInfoCopy.putAll(pxmlsInfo);
         }
 
-        final JPAIntrospection jpaIntro = JPAIntrospection.getJPAIntrospection();
-
         for (Map.Entry<String, JPAPxmlInfo> entry : pxmlsInfoCopy.entrySet()) {
             final JPAPxmlInfo value = entry.getValue();
 
-            jpaIntro.beginPXmlInfoVisit(value);
+            JPAIntrospection.beginPXmlInfoVisit(value);
             try {
                 out.println();
                 value.doIntrospect(out);
             } finally {
-                jpaIntro.endPXmlInfoVisit();
+                JPAIntrospection.endPXmlInfoVisit();
             }
         }
     }
